Improve compile times for the swizzles
Tried to simplify swizzles as much as possible. Because these structs, especially the swizzle3 and 4, are included and instantiated so many times, every extra function or member has a big impact on final compile times. As an example, vector_float_w saved around 30% compile time (from 32ms to 19ms). The other swizzles have less impact because there's fewer but still an improvement. The important changes are:
- remove swizzle<> templated functions and replace with namespaced macros
- remove union. The union is there only for the natvis, which is changed to a cast instead and works the same way
- remove identity swizzle for swizzle4 types. A simple copy is actually more performant if e.g. swizzling foo.grba = bar.grba
Double is harder to tackle because we distinguish between having wide vectors, but it's coming next
Added uintN swizzles to the natvis