Separate matrix_float_type and matrix_float
We need to be able to include lightweight classes in headers without bringing in everything else. Currently matrix_float_type includes all of vector_float but we should improve that as well. The reason for that is the transform extension. We could potentially add the minimal set of functionality it needs and strip the rest