Make sure scalar has the same alignment requirements as the SIMD versions
This makes me think whether the smaller types can have a more relaxed size and alignment requirement (e.g. in the case of NEON we can have a float2 that is a float32x2t instead of a float32x4)
This needs to be thought out well to avoid weird mismatches (e.g. different alignment between the platform that produces the data and the platform that consumes it)