Add min and max since it's what I need for my builtin types
A full replacement for these functions that follows the exact implementation provided for the vector types is necessary to be able to remove the dependency on cmath. If I provide some functions but not others, if I need e.g. exp2 I cannot bring in cmath even as a temporary workaround